requirements.txt
pynput
requests
Pillow
colorama
pyinstaller